#192 Refactor clone-material-icons script
