+++ bin/clone-material-icons.js
@@ -4,8 +4,27 @@ const fs = require('fs');
 ncp.limit = 16;
 const source = './node_modules/material-design-icons/iconfont';
 const destination = './public/iconfont';
+const inputListOfIcons = source + '/codepoints';
+const outputListOfIcons = './src/__jscash__/icons-names-list.js';
 
-function convertFile (data) {
+cloneMaterialIcons()
+  .then(generateListOfIcons)
+  .then(writeFileWithListOfIcons);
+
+function cloneMaterialIcons () {
+  return new Promise(((resolve, reject) => {
+    ncp(source, destination, error => {
+      if (error) {
+        reject(error);
+      } else {
+        resolve();
+      }
+    });
+  }));
+}
+
+function generateListOfIcons () {
+  const data = fs.readFileSync(inputListOfIcons, 'utf8');
   const linesToPrepare = data.split('\n');
   const lines = linesToPrepare
     .filter(line => line !== '')
@@ -14,26 +33,19 @@ function convertFile (data) {
       return `  '${iconName}': '${iconName}',\n`;
     });
   let fileContent = 'export const ICONS = {\n';
-
   fileContent += lines.join('');
-
   fileContent += '};';
 
-  fs.writeFile(destination + '/list.js', fileContent, function (err) {
+  return fileContent;
+}
+
+function writeFileWithListOfIcons (fileContent) {
+  fs.writeFile(outputListOfIcons, fileContent, function (err) {
     if (err) return console.log(err);
     console.log(`Material Icon List > ${destination}/list.js`);
     console.log('Material design icons has been cloned!');
   });
 }
 
-ncp(source, destination, error => {
-  if (error) {
-    return console.error(error);
-  }
-  setTimeout(() => {
-    const file = fs.readFileSync(destination + '/codepoints', 'utf8');
-    convertFile(file)
-  }, 100)
-});
